# project-1-todo-apps/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ini adalah contoh Aplikasi ToDo - Rest API.",
        "title": "ToDo Apps - Rest API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/to-do",
    "paths": {
        "/to-do": {
            "get": {
                "description": "Menampilkan semua daftar data ToDo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "to-do"
                ],
                "summary": "Menampilkan semua ToDo",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Todo"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Menambahkan satu data ToDo baru",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "to-do"
                ],
                "summary": "Menambahkan ToDo Baru",
                "parameters": [
                    {
                        "description": "Buat Todo",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Todo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Todo"
                        }
                    }
                }
            }
        },
        "/to-do/{id}": {
            "get": {
                "description": "Menampilkan sebuah data ToDo berdasarkan Id yang dimasukkan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "to-do"
                ],
                "summary": "Menampilkan sebuah data ToDo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ambil By Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Todo"
                        }
                    }
                }
            },
            "put": {
                "description": "Mengubah ToDo yang sudah ada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "to-do"
                ],
                "summary": "Mengubah ToDo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ubah ToDo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "minta data json",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Todo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Todo"
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus ToDo yang sudah ada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "to-do"
                ],
                "summary": "Menghapus ToDo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hapus ToDo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Todo"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Todo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                },
                "selesai": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}